+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add reply to long-press menu
 * Fix @-mentions
 * Replace nsec.app signup with njump.me
+* Add new messages button in rooms
 
 # 0.2.5